Rename module 'data_structures' to 'graph'
